requirements.txt
django==2.1.5
django-crispy-forms==1.7.2
django-tinymce==2.8.0
pillow==6.2.2
django-autocomplete-light==3.5.0
requests~=2.25.1
transliterate~=1.10.2